Removing background_subtractor and renaming 2
